+++ processing/threedidepth_algorithm.py
@@ -42,7 +42,7 @@ pluginPath = os.path.split(os.path.dirname(__file__))[0]
 Mode = namedtuple("Mode", ["name", "description"])
 
 
-class ProcessingParamterNetcdfNumber(QgsProcessingParameterNumber):
+class ProcessingParameterNetcdfNumber(QgsProcessingParameterNumber):
     def __init__(self, *args, parentParameterName="", optional=False, **kwargs):
         super().__init__(*args, **kwargs)
         self.parentParameterName = parentParameterName
@@ -108,6 +108,7 @@ class TimeSliderWidget(BASE, WIDGET):
         self.horizontalSlider.setMinimum(0)
         self.horizontalSlider.setMaximum(len(timestamps) - 1)
         self.timestamps = timestamps
+        self.set_lcd_value(0)  # also sets self.index
 
     def set_lcd_value(self, index: int):
         self.index = index
@@ -179,7 +180,7 @@ class CheckboxTimeSliderWidget(TimeSliderWidget, WIDGET, BASE):
 
 class ThreediDepth(QgsProcessingAlgorithm):
     """
-    Calculates waterdepths for 3Di results
+    Calculates water depth or water level rasters from 3Di result NetCDF
     """
 
     # Constants used to refer to parameters and outputs. They will be
@@ -220,7 +221,7 @@ class ThreediDepth(QgsProcessingAlgorithm):
         Returns the translated algorithm name, which should be used for any
         user-visible display of the algorithm name.
         """
-        return self.tr("Water depth")
+        return self.tr("Water depth or water level raster")
 
     def group(self):
         """Returns the name of the group this algorithm belongs to"""
@@ -232,7 +233,7 @@ class ThreediDepth(QgsProcessingAlgorithm):
 
     def shortHelpString(self):
         """Returns a localised short helper string for the algorithm"""
-        return self.tr("Calculate water depths for 3Di results.")
+        return self.tr("Calculate water depth or water level raster for specified timestep")
 
     def initAlgorithm(self, config=None):
         """Here we define the inputs and output of the algorithm"""
@@ -259,7 +260,7 @@ class ThreediDepth(QgsProcessingAlgorithm):
             )
         )
         self.addParameter(
-            ProcessingParamterNetcdfNumber(
+            ProcessingParameterNetcdfNumber(
                 name=self.CALCULATION_STEP_INPUT,
                 description=self.tr(
                     "The timestep in the simulation for which you want to generate a raster"
@@ -269,7 +270,7 @@ class ThreediDepth(QgsProcessingAlgorithm):
             )
         )
         self.addParameter(
-            ProcessingParamterNetcdfNumber(
+            ProcessingParameterNetcdfNumber(
                 name=self.CALCULATION_STEP_END_INPUT,
                 description=self.tr(
                     "In case you want to export water depths of multiple timesteps, enable this option and select "
@@ -290,17 +291,20 @@ class ThreediDepth(QgsProcessingAlgorithm):
 
         self.addParameter(
             QgsProcessingParameterRasterDestination(
-                self.WATER_DEPTH_OUTPUT, self.tr("Water depth raster")
+                self.WATER_DEPTH_OUTPUT, self.tr('Water depth or water level raster')
             )
         )
 
     def processAlgorithm(self, parameters, context, feedback):
         """
-        Create the waterdepth raster with the provided user inputs
+        Create the water depth raster with the provided user inputs
         """
         waterdepth_output_file = self.parameterAsOutputLayer(
             parameters, self.WATER_DEPTH_OUTPUT, context
         )
+        dem_filename = self.parameterAsRasterLayer(
+            parameters, self.DEM_INPUT, context
+        ).source()
         mode_index = self.parameterAsEnum(parameters, self.MODE_INPUT, context)
 
         endstep = parameters[self.CALCULATION_STEP_END_INPUT]
@@ -319,7 +323,7 @@ class ThreediDepth(QgsProcessingAlgorithm):
             calculate_waterdepth(
                 gridadmin_path=parameters[self.GRIDADMIN_INPUT],
                 results_3di_path=parameters[self.RESULTS_3DI_INPUT],
-                dem_path=parameters[self.DEM_INPUT],
+                dem_path=dem_filename,
                 waterdepth_path=waterdepth_output_file,
                 calculation_steps=timesteps,
                 mode=self.MODES[mode_index].name,